Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

If it's breaking something, it should go.

I know some would prefer that the broken patches be fixed instead of reverted - but considering we now have bugs in RC releases that are SUPPOSED to be stable due to all this stuff being thrown in at once without any testing at all, it should be yanked until it can be resubmitted and proven not to break anything.

Patch-set: 1