Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

Is there a reason we are reverting patches wholesale instead of fixing the problems with them?

The problem appears to be that wilhelm/src/itf/IAndroidEffectCapabilities.c was originally assigning a pointer to .type, and is now assigning the contents of it.

Patch-set: 1