Update patch set 1
Patch Set 1:

Christopher:

Here's my philosophy on it:

If you submit code, and it breaks things, it is YOUR responsibility to fix it, not someone else's.  If you break things, you shouldn't complain that your patch got reverted.

Revert the broken crud, and let whoever wants the code to be in fix it and resubmit.

Patch-set: 1