Update patch set 1
Patch Set 1:

Except this was a patch we actively pulled from an upstream project. We take a lot from those projects, and it's only fair to give something back; it's as much a two-way street as we can make it.

If you want to maintain a passive parasitic approach to upstreams and/or wait for stuff to magically fix itself, CM isn't quite the place where you want to be.

Patch-set: 1