Update patch set 1
Patch Set 1:

While I'd usually agree with you I can't in this specific case because the problematic code is part from a parser.

Unless you want to investigate time to analyze the full memory dump, a parse stacktrace is not useful in most cases because u have to know what went wrong in the document.

But I agree with you that it should still fail, so maybe using abort() would the best way?

Patch-set: 1