Update patch set 4
Patch Set 4:

There's still stuff that's defined in here that gets called by the HALs (AMR_WB_FRAMESIZE was the one I ran into on 8994) IDK where they put it now, but yeah, it's not here. 8916 I ran into AUDIO_FORMAT_DTS_LBR, so that needs it too. I'm definitely not pushing for this to be merged, but if we get rid of it, we need to find where these definitions moved to.

Patch-set: 4